Weighted DNS routing policy
Updates the route53 DNS record creation to use a weigthed routing policy
type instead of simple. All policy types except simple require a single
target value per record, so a dens record can now have multiple
endpoints listed as part of ot's spec, one for each ip address listed in
an ingress.

The approach implemented is very similar to how external-dns does
things, this should hopefully make it easier to switch to using
exernal-dns in the future if we decide to do so.

Note: This breaks the deletion of old records when an IP is removed from
an ingress.