Add endpoints to DNSZoneStatus
DNSZoneStatus.Endpoints are the last endpoints that were successfully
published to the provider.  Provides a simple mechanism to store the
current provider records in order to delete any that are no longer
present in DNSRecordSpec.Endpoints during reconciliation.

Note: This will no be required if/when we switch to using external-dns
since when running with a "sync" policy it will clean up unused records
automatically.