Add deploy job to GH workflows
Adds a new deploy job to the "Build and Publish KCP GLBC Image" workflow
that will deploy the new image to a cluster in a given environment.
Currently there are 2 envs configured "stable" and "unstable" with
updates to main being deployed to "unstable". When a new release is
created e.g. release-1.0.0, the workflow shoud be updated in that branch
to point to "stable" in order to initiate deployments of the release
branch to the "stable" environment.

kustomize manifest updates:

* Remove namePrefix, prepend all resource names with kcp-glbc- (Casues
  issues when re-deploying and naming screts/config)
* Move namespace creation to default
* Move manager config creation to it's own kustomization
* Remove crds from deployments (They aren't needed for a deployment of
  the controller, they need to be applied to the kcp cluster)
* Use recreate strategy in deployment
* Update manager role (Add what's needed for the glbc kubeconfig)