Update kustomize directory structure
Updates the default directory structure for a deployment of kcp-glbc and
moves the kcp resources required to run glbc and cert-manager in kcp
into the kustomization of a glbc deployment. Deploy script updated to
generate all deployment resources through kustomize with the excpetion
of APIBindings and APIExports which can be dealt with at a later stage.

Output sync target resources to sync-targets directory of deployment.
Output apiexports and apibindings into kcp-glbc/apiexports directory of
deployment.