Add ingress deletion to test cleanup
Adds the deletion of all ingress resources to the testing cleanup phase.
When testing on the kcp service we need to ensure that resources we
create inside the ws and ns are deleted before we delete the ns and
ultimately the ws or it can end up in a stuck deleting phase.